requirements.txt
python-binance
numpy
websocket_client
